Fixes for refset in 19 version bundle.

+++ src/test/java/org/ihtsdo/rules/test/context/TestContexts.java
@@ -232,7 +232,7 @@ public class TestContexts extends TestCase {
 			config = tf.newAceFrameConfig();
 			config.addViewPosition(tf.newPosition(
 					tf.getPath(new UUID[] {UUID.fromString("2faa9260-8fb2-11db-b606-0800200c9a66")}), 
-					Integer.MAX_VALUE));
+					Long.MAX_VALUE));
 			config.addEditingPath( tf.getPath(new UUID[] {UUID.fromString("2faa9260-8fb2-11db-b606-0800200c9a66")}));
 			config.addPromotionPath( tf.getPath(new UUID[] {UUID.fromString("2faa9260-8fb2-11db-b606-0800200c9a66")}));
